full deprecate str collect
