+++ crates/nu-command/src/deprecated/deprecated_commands.rs
@@ -23,5 +23,6 @@ pub fn deprecated_commands() -> HashMap<String, String> {
         ("str lpad".to_string(), "fill".to_string()),
         ("str rpad".to_string(), "fill".to_string()),
         ("benchmark".to_string(), "timeit".to_string()),
+        ("str collect".to_string(), "str join".to_string()),
     ])
 }
